Conflicts:
	base/locale/en/LC_MESSAGES/django.po
	base/locale/fr_BE/LC_MESSAGES/django.po

+++ dissertation/views/proposition_dissertation.py
@@ -23,7 +23,6 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-from dissertation.utils.request import redirect_if_none
 from django.shortcuts import redirect
 from django.contrib.auth.decorators import login_required
 from base import models as mdl
@@ -40,7 +39,6 @@ from django.contrib.auth.decorators import user_passes_test
 from base.views import layout
 import time
 from django.http import HttpResponse
-from django.core.exceptions import ObjectDoesNotExist
 
 
 def detect_in_request(request, wanted_key, wanted_value):
@@ -99,7 +97,8 @@ def manager_proposition_dissertations(request):
 @user_passes_test(adviser.is_manager)
 def manager_proposition_dissertation_delete(request, pk):
     proposition = proposition_dissertation.find_by_id(pk)
-    redirect_if_none(proposition,'manager_proposition_dissertations')
+    if proposition is None:
+        return redirect('manager_proposition_dissertations')
     proposition.deactivate()
     return redirect('manager_proposition_dissertations')
 
@@ -108,7 +107,8 @@ def manager_proposition_dissertation_delete(request, pk):
 @user_passes_test(adviser.is_manager)
 def manager_proposition_dissertation_detail(request, pk):
     proposition = proposition_dissertation.find_by_id(pk)
-    redirect_if_none(proposition, 'manager_proposition_dissertations')
+    if proposition is None:
+        return redirect('manager_proposition_dissertations')
     offer_propositions = proposition_offer.find_by_proposition_dissertation(proposition)
     person = mdl.person.find_by_user(request.user)
     adv = adviser.search_by_person(person)
@@ -137,7 +137,8 @@ def manager_proposition_dissertation_detail(request, pk):
 @user_passes_test(adviser.is_manager)
 def manage_proposition_dissertation_edit(request, pk):
     proposition = proposition_dissertation.find_by_id(pk)
-    redirect_if_none(proposition, 'manager_proposition_dissertations')
+    if proposition is None:
+        return redirect('manager_proposition_dissertations')
     offer_propositions = offer_proposition.find_all_ordered_by_acronym()
     offer_propositions_group = offer_proposition_group.find_all_ordered_by_name_short()
     offer_propositions_error = None
@@ -162,14 +163,15 @@ def manage_proposition_dissertation_edit(request, pk):
                           'offer_propositions': offer_propositions,
                           'offer_propositions_error': offer_propositions_error,
                           'proposition_offers': proposition_offers,
-                          'offer_proposition_group':offer_propositions_group})
+                          'offer_proposition_group': offer_propositions_group})
 
 
 @login_required
 @user_passes_test(adviser.is_manager)
 def manager_proposition_dissertations_jury_edit(request, pk):
     prop_role = proposition_role.get_by_id(pk)
-    redirect_if_none(proposition_role, 'manager_proposition_dissertations')
+    if prop_role is None:
+        return redirect('manager_proposition_dissertations')
     proposition = prop_role.proposition_dissertation
     return redirect('manager_proposition_dissertation_detail', pk=proposition.pk)
 
@@ -178,7 +180,8 @@ def manager_proposition_dissertations_jury_edit(request, pk):
 @user_passes_test(adviser.is_manager)
 def manager_proposition_dissertations_jury_new(request, pk):
     proposition = proposition_dissertation.find_by_id(pk)
-    redirect_if_none(proposition, 'manager_proposition_dissertations')
+    if proposition is None:
+        return redirect('manager_proposition_dissertations')
     count_proposition_role = PropositionRole.objects.filter(proposition_dissertation=proposition).count()
     if request.method == "POST":
         form = ManagerPropositionRoleForm(request.POST)
@@ -205,7 +208,8 @@ def manager_proposition_dissertations_jury_new(request, pk):
 @user_passes_test(adviser.is_manager)
 def manager_proposition_dissertations_role_delete(request, pk):
     prop_role = proposition_role.get_by_id(pk)
-    redirect_if_none(prop_role, 'manager_proposition_dissertations')
+    if prop_role is None:
+        return redirect('manager_proposition_dissertations')
     proposition = prop_role.proposition_dissertation
     prop_role.delete()
     return redirect('manager_proposition_dissertation_detail', pk=proposition.pk)
@@ -303,7 +307,8 @@ def proposition_dissertations(request):
 @user_passes_test(adviser.is_teacher)
 def proposition_dissertation_delete(request, pk):
     proposition = proposition_dissertation.find_by_id(pk)
-    redirect_if_none(proposition, 'proposition_dissertations')
+    if proposition is None:
+        return redirect('proposition_dissertations')
     proposition.deactivate()
     return redirect('proposition_dissertations')
 
@@ -312,7 +317,8 @@ def proposition_dissertation_delete(request, pk):
 @user_passes_test(adviser.is_teacher)
 def proposition_dissertation_detail(request, pk):
     proposition = proposition_dissertation.find_by_id(pk)
-    redirect_if_none(proposition, 'proposition_dissertations')
+    if proposition is None:
+        return redirect('proposition_dissertations')
     offer_propositions = proposition_offer.find_by_proposition_dissertation(proposition)
     count_use = dissertation.count_by_proposition(proposition)
     percent = count_use * 100 / proposition.max_number_student if proposition.max_number_student else 0
@@ -339,7 +345,8 @@ def proposition_dissertation_detail(request, pk):
 @user_passes_test(adviser.is_teacher)
 def proposition_dissertation_edit(request, pk):
     proposition = proposition_dissertation.find_by_id(pk)
-    redirect_if_none(proposition, 'proposition_dissertations')
+    if proposition is None:
+        return redirect('proposition_dissertations')
     adv = get_current_adviser(request)
     offer_propositions = offer_proposition.find_all_ordered_by_acronym()
     offer_propositions_group = offer_proposition_group.find_all_ordered_by_name_short()
@@ -429,7 +436,8 @@ def proposition_dissertations_search(request):
 @user_passes_test(adviser.is_teacher)
 def proposition_dissertations_jury_edit(request, pk):
     prop_role = proposition_role.get_by_id(pk)
-    redirect_if_none(prop_role, 'proposition_dissertations')
+    if prop_role is None:
+        return redirect('proposition_dissertations')
     proposition = prop_role.proposition_dissertation
     return redirect('proposition_dissertation_detail', pk=proposition.pk)
 
@@ -438,7 +446,8 @@ def proposition_dissertations_jury_edit(request, pk):
 @user_passes_test(adviser.is_teacher)
 def proposition_dissertations_jury_new(request, pk):
     proposition = proposition_dissertation.find_by_id(pk)
-    redirect_if_none(proposition, 'proposition_dissertations')
+    if proposition is None:
+        return redirect('proposition_dissertations')
     count_proposition_role = PropositionRole.objects.filter(proposition_dissertation=proposition).count()
     adv = get_current_adviser(request)
     if proposition.author == adv or proposition.creator == adv.person:
@@ -469,7 +478,8 @@ def proposition_dissertations_jury_new(request, pk):
 @user_passes_test(adviser.is_teacher)
 def proposition_dissertations_role_delete(request, pk):
     prop_role = proposition_role.get_by_id(pk)
-    redirect_if_none(prop_role, 'proposition_dissertations')
+    if prop_role is None:
+        return redirect('proposition_dissertations')
     proposition = prop_role.proposition_dissertation
     adv = get_current_adviser(request)
